model-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-multipart==0.0.9
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-redis==0.46b0
opentelemetry-exporter-otlp-proto-http==1.25.0
opencv-python==4.9.0.80
numpy==1.26.4
pyyaml==6.0.1
torch==2.5.1
